+++ packages/jagos/ScriptWrapperMain.ts
@@ -1,9 +1,9 @@
-import { makeJagoOnError } from "^bee-common";
+import { makeBeeOnError } from "^bee-common";
 import { makeSend, registerRejectionHandlers, wppMain } from "^jab-node";
 
 //register rejection handlers
 
-registerRejectionHandlers(makeJagoOnError(makeSend()));
+registerRejectionHandlers(makeBeeOnError(makeSend()));
 
 //process preloader