+++ packages/jagov/onServerMessage.ts
@@ -1,7 +1,7 @@
 import { assertNever, tos, tryPropString } from "^jab";
 import { ConsoleEntry } from "^console";
 import { ServerMessage } from "^jagoc";
-import { JagoLogEntry } from "^bee-common";
+import { BeeLogEntry } from "^bee-common";
 import { StateCallbacks } from ".";
 
 type Deps = Pick<StateCallbacks, "setProcessStatus"> & {
@@ -68,7 +68,7 @@ export const handleScriptMessage = (
     return;
   }
 
-  const msg = umsg as JagoLogEntry;
+  const msg = umsg as BeeLogEntry;
 
   switch (msg.type) {
     case "log":