+++ packages/tests/all/jago mainProvToJago 1.ja.ts
@@ -1,9 +1,9 @@
-import { mainProvToJago } from "^bee-common";
+import { makeMainBeeProv } from "^bee-common";
 import { TestProvision } from "^jarun";
 import { makeJagoSend } from "^tests/_fixture";
 
 export default async (prov: TestProvision) => {
-  const mainProv = mainProvToJago(makeJagoSend(prov));
+  const mainProv = makeMainBeeProv(makeJagoSend(prov));
 
   mainProv.onError(new Error("hello"));
   mainProv.onError(new Error("there's more"), ["here it is"]);