+++ packages/tests/all/jago mainProvToJago 2.ja.ts
@@ -1,11 +1,11 @@
-import { mainProvToJago } from "^bee-common";
+import { makeMainBeeProv } from "^bee-common";
 import { TestProvision } from "^jarun";
 import { makeJagoSend } from "^tests/_fixture";
 
 //stream has prefix
 
 export default async (prov: TestProvision) => {
-  const mainProv = mainProvToJago(makeJagoSend(prov), "prefix.");
+  const mainProv = makeMainBeeProv(makeJagoSend(prov), "prefix.");
 
   mainProv.logStream("the name", "hello");
 };